requirements.txt
cadquery-ocp==7.8.1.1.post1
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pillow==11.3.0
pyparsing==3.2.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
six==1.17.0
vtk==9.3.1
